% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/audio.R
\name{audio_duration}
\alias{audio_duration}
\title{Get Audio Duration}
\usage{
audio_duration(file)
}
\arguments{
\item{file}{Path to audio file}
}
\value{
Duration in seconds
}
\description{
Get Audio Duration
}
