% tinyrox says don't edit this manually, but it can't stop you!
\name{compression_ratio}
\alias{compression_ratio}
\title{Compression Ratio}
\usage{
compression_ratio(text)
}
\arguments{
\item{text}{Character string}
}
\value{
Numeric compression ratio
}
\description{
Ratio of raw to compressed text size. High values indicate repetitive
or hallucinated output.
}
