% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/decoder.R
\name{create_decoder}
\alias{create_decoder}
\title{Create Decoder from Config}
\usage{
create_decoder(config)
}
\arguments{
\item{config}{Model configuration from whisper_config()}
}
\value{
WhisperDecoder module
}
\description{
Create Decoder from Config
}
