% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/audio.R
\name{create_mel_filterbank}
\alias{create_mel_filterbank}
\title{Create Mel Filterbank}
\usage{
create_mel_filterbank(
  n_fft = WHISPER_N_FFT,
  n_mels = 80L,
  sample_rate = WHISPER_SAMPLE_RATE
)
}
\arguments{
\item{n_fft}{FFT size}

\item{n_mels}{Number of mel bins}

\item{sample_rate}{Audio sample rate}
}
\value{
Mel filterbank matrix (n_mels x n_freqs)
}
\description{
Create a mel filterbank matrix for converting STFT to mel spectrogram.
}
