% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{decode_timestamp}
\alias{decode_timestamp}
\title{Decode Timestamp Token}
\usage{
decode_timestamp(token_id, model = "tiny")
}
\arguments{
\item{token_id}{Token ID}

\item{model}{Model name for correct token IDs}
}
\value{
Time in seconds
}
\description{
Decode Timestamp Token
}
