% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{.model_sizes}
\alias{.model_sizes}
\title{Model Download Utilities}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
.model_sizes
}
\description{
Download Whisper models from HuggingFace using hfhub.
}
\keyword{datasets}
