% tinyrox says don't edit this manually, but it can't stop you!
\name{download_tokenizer_files}
\alias{download_tokenizer_files}
\title{Download Tokenizer Files from HuggingFace}
\usage{
download_tokenizer_files(model, revision = NULL)
}
\arguments{
\item{model}{Model name}

\item{revision}{Optional exact 40-hex commit to resolve against.}
}
\description{
Download Tokenizer Files from HuggingFace
}
