% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_html_response_url}
\alias{get_html_response_url}
\title{Query the response header of a webpage}
\usage{
get_html_response_url(urls)
}
\arguments{
\item{urls}{A character vector of URLs to iterate over}
}
\description{
Query the response header of a webpage
}
\examples{
 get_html_response_url(c('http://nytimes.com', 'http://nytimes.com/es'))

}
