% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_language_url}
\alias{get_language_url}
\title{Determine language of website}
\usage{
get_language_url(urls, show_url = TRUE, show_progress = TRUE,
  wait_time = 5)
}
\arguments{
\item{urls}{A character vector of URLs to iterate over}

\item{show_url}{Print URL of current iteration}

\item{show_progress}{Show overall progress periodically}

\item{wait_time}{Time to wait before moving from bad URL}
}
\description{
Determine language of website
}
\examples{
 get_language_url(c('http://nytimes.com', 'http://nytimes.com/es'))

}
