% tinyrox says don't edit this manually, but it can't stop you!
\name{get_model_path}
\alias{get_model_path}
\title{Get Model Cache Path}
\usage{
get_model_path(model, revision = NULL)
}
\arguments{
\item{model}{Model name}

\item{revision}{Optional exact 40-hex commit to resolve against.}
}
\value{
Path to model directory in hfhub cache
}
\description{
Get Model Cache Path
}
