% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences_url.R
\name{get_sentences_url}
\alias{get_sentences_url}
\title{Get sentences from a part of a website}
\usage{
get_sentences_url(url, node, process = TRUE)
}
\arguments{
\item{url}{URL to retrieve sentences from}

\item{node}{Part of webpage to extract text from}

\item{process}{Also try to remove non-language text and extract sentences}
}
\description{
Get sentences from a part of a website
}
\examples{
 get_sentences_url('http://nytimes.com/', 'body', process=FALSE)

}
