% tinyrox says don't edit this manually, but it can't stop you!
\name{get_weights_path}
\alias{get_weights_path}
\title{Get Path to Model Weights}
\usage{
get_weights_path(model, revision = NULL)
}
\arguments{
\item{model}{Model name}

\item{revision}{Optional exact 40-hex commit to resolve against.}
}
\value{
Path to safetensors file
}
\description{
Get Path to Model Weights
}
