% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{is_timestamp_token}
\alias{is_timestamp_token}
\title{Check if Token is Timestamp}
\usage{
is_timestamp_token(token_id, model = "tiny")
}
\arguments{
\item{token_id}{Token ID}

\item{model}{Model name for correct token IDs}
}
\value{
TRUE if timestamp token
}
\description{
Check if Token is Timestamp
}
