% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/model.R
\name{load_decoder_weights}
\alias{load_decoder_weights}
\title{Load Decoder Weights}
\usage{
load_decoder_weights(decoder, weights)
}
\arguments{
\item{decoder}{WhisperDecoder module}

\item{weights}{Named list of tensors}
}
\description{
Load Decoder Weights
}
