% tinyrox says don't edit this manually, but it can't stop you!
\name{load_encoder_weights}
\alias{load_encoder_weights}
\title{Load Encoder Weights}
\usage{
load_encoder_weights(encoder, weights)
}
\arguments{
\item{encoder}{WhisperEncoder module}

\item{weights}{Named list of tensors}
}
\description{
Load Encoder Weights
}
