% tinyrox says don't edit this manually, but it can't stop you!
\name{load_whisper_weights}
\alias{load_whisper_weights}
\title{Load Weights from Safetensors}
\usage{
load_whisper_weights(model, weights_path, verbose = TRUE)
}
\arguments{
\item{model}{WhisperModel module}

\item{weights_path}{Path to safetensors file}

\item{verbose}{Print loading messages}
}
\description{
Load Weights from Safetensors
}
