% tinyrox says don't edit this manually, but it can't stop you!
\name{medfilt1}
\alias{medfilt1}
\title{1D Median Filter}
\usage{
medfilt1(x, width = 7L)
}
\arguments{
\item{x}{Numeric vector}

\item{width}{Filter width (must be odd)}
}
\value{
Filtered numeric vector of same length
}
\description{
Apply a sliding median filter to a numeric vector.
}
