% tinyrox says don't edit this manually, but it can't stop you!
\name{pad_or_trim}
\alias{pad_or_trim}
\title{Pad or Trim Audio to Fixed Length}
\usage{
pad_or_trim(audio, length = WHISPER_N_SAMPLES)
}
\arguments{
\item{audio}{Numeric vector of audio samples}

\item{length}{Target length in samples (default: 30s at 16kHz)}
}
\value{
Numeric vector of specified length
}
\description{
Pad or Trim Audio to Fixed Length
}
