% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/devices.R
\name{parse_device}
\alias{parse_device}
\title{Parse Device Argument}
\usage{
parse_device(device = "auto")
}
\arguments{
\item{device}{Character or torch device. "auto" uses GPU if available.}
}
\value{
torch device object
}
\description{
Parse Device Argument
}
