% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_html_text}
\alias{parse_html_text}
\title{Parse the text extracted from a webpage}
\usage{
parse_html_text(text)
}
\arguments{
\item{text}{URL to retrieve sentences from}
}
\description{
Parse the text extracted from a webpage
}
\examples{
parse_html_text(get_sentences_url(url='http://nytimes.com/es', node='body', process=FALSE))
}
