% tinyrox says don't edit this manually, but it can't stop you!
\name{print.whisper_resident}
\alias{print.whisper_resident}
\title{Print a Resident Handle}
\usage{
\method{print}{whisper_resident}(x, ...)
}
\arguments{
\item{x}{A `whisper_resident` handle.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a Resident Handle
}
