% tinyrox says don't edit this manually, but it can't stop you!
\name{print.whisper_result}
\alias{print.whisper_result}
\title{Print a Transcription Result}
\usage{
\method{print}{whisper_result}(x, ...)
}
\arguments{
\item{x}{A `transcribe()` result.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a Transcription Result
}
