% tinyrox says don't edit this manually, but it can't stop you!
\name{split_audio}
\alias{split_audio}
\title{Split Long Audio into Chunks}
\usage{
split_audio(file, chunk_length = 30, overlap = 1)
}
\arguments{
\item{file}{Path to audio file}

\item{chunk_length}{Chunk length in seconds}

\item{overlap}{Overlap between chunks in seconds}
}
\value{
List of audio chunks (numeric vectors)
}
\description{
Split audio longer than 30 seconds into overlapping chunks.
}
