% tinyrox says don't edit this manually, but it can't stop you!
\name{tokenizer_decode}
\alias{tokenizer_decode}
\title{Decode Token IDs to Text}
\usage{
tokenizer_decode(ids, id_to_token, special_tokens)
}
\arguments{
\item{ids}{Integer vector of token IDs}

\item{id_to_token}{Mapping from ID to token}

\item{special_tokens}{Special token info}
}
\value{
Character string
}
\description{
Decode Token IDs to Text
}
