% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_plot.R
\name{wf_plot}
\alias{wf_plot}
\title{Plot Wright-Fisher Simulations}
\usage{
wf_plot(num_sims = 100L, show_hist = FALSE, ...)
}
\arguments{
\item{num_sims}{Number of simulations to run.}

\item{show_hist}{Show distribution of allele frequencies at end of
simulations.}

\item{...}{Parameters passed to \code{\link{wf_sim}} function.}
}
\value{
ggplot showing allele frequencies over generations.
}
\description{
Generates a plot of Wright-Fisher simulations with specified parameters.
}
\details{
Generates a plot of Wright-Fisher simulations using
provided parameters. It uses the \code{\link{wf_sim}} function to simulate
genetic drift across generations and plots allele trajectory.
}
\examples{
wf_plot()
wf_plot(p = 1/250, fit_aa = .8)
wf_plot(p = 1/250, mut_to = .05, mut_from = .03)

}
