% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_shiny.R
\name{wf_shiny}
\alias{wf_shiny}
\title{Run interactive Wright-Fisher Genetic Drift Simulation}
\usage{
wf_shiny()
}
\value{
Shiny app to interactively explore genetic drift simulations.
}
\description{
Run interactive Wright-Fisher Genetic Drift Simulation
}
\examples{
\dontrun{
wf_shiny()
}

}
\seealso{
\code{\link{wf_plot}} for generating Wright-Fisher simulation plots.
}
