% tinyrox says don't edit this manually, but it can't stop you!
\name{whisper_attention}
\alias{whisper_attention}
\title{Multi-Head Self-Attention}
\usage{
whisper_attention(n_state, n_head)
}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of attention heads}
}
\description{
Multi-Head Self-Attention
}
