% tinyrox says don't edit this manually, but it can't stop you!
\name{whisper_decoder_layer}
\alias{whisper_decoder_layer}
\title{Decoder Layer}
\usage{
whisper_decoder_layer(n_state, n_head)
}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of attention heads}
}
\description{
Pre-norm transformer decoder layer with self-attention and cross-attention.
}
