% tinyrox says don't edit this manually, but it can't stop you!
\name{whisper_encoder_layer}
\alias{whisper_encoder_layer}
\title{Encoder Layer}
\usage{
whisper_encoder_layer(n_state, n_head)
}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of attention heads}
}
\description{
Pre-norm transformer encoder layer.
}
