% tinyrox says don't edit this manually, but it can't stop you!
\name{whisper_lang_from_id}
\alias{whisper_lang_from_id}
\title{Get Language Code from Token ID}
\usage{
whisper_lang_from_id(token_id)
}
\arguments{
\item{token_id}{Integer token ID (e.g., 50259 for English)}
}
\value{
Two-letter language code
}
\description{
Reverse lookup: convert a language token ID back to a two-letter code.
}
