% tinyrox says don't edit this manually, but it can't stop you!
\name{whisper_language_table}
\alias{whisper_language_table}
\title{Whisper Language Table}
\usage{
whisper_language_table()
}
\value{
Named integer vector (language code -> offset from 50259)
}
\description{
Returns the named integer vector mapping language codes to offsets.
}
