% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/model.R
\name{whisper_model}
\alias{whisper_model}
\title{Whisper Model}
\usage{
whisper_model(config)
}
\arguments{
\item{config}{Model configuration}
}
\description{
Full Whisper model combining encoder and decoder.
Whisper Model Module
}
