% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{WHISPER_SAMPLE_RATE}
\alias{WHISPER_SAMPLE_RATE}
\title{Whisper Audio Constants}
\format{
An object of class \code{integer} of length 1.
}
\usage{
WHISPER_SAMPLE_RATE
}
\description{
Whisper Audio Constants
}
\keyword{datasets}
