% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{apply_bpe}
\alias{apply_bpe}
\title{Apply BPE Merges}
\usage{
apply_bpe(tokens, merge_ranks)
}
\arguments{
\item{tokens}{Character vector of tokens}

\item{merge_ranks}{Named vector of merge rankings}
}
\value{
Character vector after BPE merges
}
\description{
Apply BPE Merges
}
