% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/model.R
\name{copy_if_exists}
\alias{copy_if_exists}
\title{Copy Weight if Exists}
\usage{
copy_if_exists(param, weights, name)
}
\arguments{
\item{param}{Target parameter}

\item{weights}{Weight dictionary}

\item{name}{Weight name}
}
\description{
Copy Weight if Exists
}
