% tinyrox says don't edit this manually, but it can't stop you!
\name{create_decoder}
\alias{create_decoder}
\title{Create Decoder from Config}
\usage{
create_decoder(config)
}
\arguments{
\item{config}{Model configuration from whisper_config()}
}
\value{
WhisperDecoder module
}
\description{
Create Decoder from Config
}
