% tinyrox says don't edit this manually, but it can't stop you!
\name{create_encoder}
\alias{create_encoder}
\title{Create Encoder from Config}
\usage{
create_encoder(config)
}
\arguments{
\item{config}{Model configuration from whisper_config()}
}
\value{
WhisperEncoder module
}
\description{
Create Encoder from Config
}
