% tinyrox says don't edit this manually, but it can't stop you!
\name{decode_bpe_bytes}
\alias{decode_bpe_bytes}
\title{Decode BPE Bytes Back to Text}
\usage{
decode_bpe_bytes(text)
}
\arguments{
\item{text}{Text with BPE byte tokens}
}
\value{
Decoded text
}
\description{
Decode BPE Bytes Back to Text
}
