% tinyrox says don't edit this manually, but it can't stop you!
\name{decode_timestamp}
\alias{decode_timestamp}
\title{Decode Timestamp Token}
\usage{
decode_timestamp(token_id, model = "tiny")
}
\arguments{
\item{token_id}{Token ID}

\item{model}{Model name for correct token IDs}
}
\value{
Time in seconds
}
\description{
Decode Timestamp Token
}
