% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{download_tokenizer_files}
\alias{download_tokenizer_files}
\title{Download Tokenizer Files from HuggingFace}
\usage{
download_tokenizer_files(model)
}
\arguments{
\item{model}{Model name}
}
\description{
Download Tokenizer Files from HuggingFace
}
