% tinyrox says don't edit this manually, but it can't stop you!
\name{download_tokenizer_files}
\alias{download_tokenizer_files}
\title{Download Tokenizer Files from HuggingFace}
\usage{
download_tokenizer_files(model)
}
\arguments{
\item{model}{Model name}
}
\description{
Download Tokenizer Files from HuggingFace
}
