% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/download.R
\name{download_whisper_model}
\alias{download_whisper_model}
\title{Download Model from HuggingFace}
\usage{
download_whisper_model(model = "tiny", force = FALSE)
}
\arguments{
\item{model}{Model name}

\item{force}{Re-download even if exists}
}
\value{
Path to model directory
}
\description{
Download Model from HuggingFace
}
