% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{ensure_tokenizer_files}
\alias{ensure_tokenizer_files}
\title{Ensure Tokenizer Files are Downloaded}
\usage{
ensure_tokenizer_files(model)
}
\arguments{
\item{model}{Model name}
}
\value{
Path to vocab directory (directory containing vocab.json)
}
\description{
Ensure Tokenizer Files are Downloaded
}
