% tinyrox says don't edit this manually, but it can't stop you!
\name{ensure_tokenizer_files}
\alias{ensure_tokenizer_files}
\title{Ensure Tokenizer Files are Downloaded}
\usage{
ensure_tokenizer_files(model)
}
\arguments{
\item{model}{Model name}
}
\value{
Path to vocab directory (directory containing vocab.json)
}
\description{
Ensure Tokenizer Files are Downloaded
}
