% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/download.R
\name{get_model_path}
\alias{get_model_path}
\title{Get Model Cache Path}
\usage{
get_model_path(model)
}
\arguments{
\item{model}{Model name}
}
\value{
Path to model directory in hfhub cache
}
\description{
Get Model Cache Path
}
