% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/download.R
\name{get_model_path}
\alias{get_model_path}
\title{Model Download Utilities}
\usage{
get_model_path(model)
}
\arguments{
\item{model}{Model name}
}
\value{
Path to model directory in hfhub cache
}
\description{
Download Whisper models from HuggingFace using hfhub. Get Model Cache
Path
}
