% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/download.R
\name{get_weights_path}
\alias{get_weights_path}
\title{Get Path to Model Weights}
\usage{
get_weights_path(model)
}
\arguments{
\item{model}{Model name}
}
\value{
Path to safetensors file
}
\description{
Get Path to Model Weights
}
