% tinyrox says don't edit this manually, but it can't stop you!
\name{list_downloaded_models}
\alias{list_downloaded_models}
\title{List Downloaded Models}
\usage{
list_downloaded_models()
}
\value{
Character vector of downloaded model names
}
\description{
List Downloaded Models
}
\examples{
list_downloaded_models()
}
