% tinyrox says don't edit this manually, but it can't stop you!
\name{load_added_tokens}
\alias{load_added_tokens}
\title{Load Added Tokens from HuggingFace}
\usage{
load_added_tokens(repo)
}
\arguments{
\item{repo}{HuggingFace repo ID}
}
\value{
Named list of token -> ID mappings, or NULL if not found
}
\description{
Load Added Tokens from HuggingFace
}
