% tinyrox says don't edit this manually, but it can't stop you!
\name{load_decoder_weights}
\alias{load_decoder_weights}
\title{Load Decoder Weights}
\usage{
load_decoder_weights(decoder, weights)
}
\arguments{
\item{decoder}{WhisperDecoder module}

\item{weights}{Named list of tensors}
}
\description{
Load Decoder Weights
}
