% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/model.R
\name{load_encoder_weights}
\alias{load_encoder_weights}
\title{Load Encoder Weights}
\usage{
load_encoder_weights(encoder, weights)
}
\arguments{
\item{encoder}{WhisperEncoder module}

\item{weights}{Named list of tensors}
}
\description{
Load Encoder Weights
}
