% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/download.R
\name{model_exists}
\alias{model_exists}
\title{Check if Model is Downloaded}
\usage{
model_exists(model)
}
\arguments{
\item{model}{Model name}
}
\value{
TRUE if model weights exist locally
}
\description{
Check if Model is Downloaded
}
