% tinyrox says don't edit this manually, but it can't stop you!
\name{parse_device}
\alias{parse_device}
\title{Parse Device Argument}
\usage{
parse_device(device = "auto")
}
\arguments{
\item{device}{Character or torch device. "auto" uses GPU if available.}
}
\value{
torch device object
}
\description{
Parse Device Argument
}
