% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/encoder.R
\name{whisper_attention}
\alias{whisper_attention}
\title{Whisper Encoder}
\usage{
whisper_attention(n_state, n_head)
}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of attention heads}
}
\description{
Transformer encoder for processing mel spectrograms. Multi-Head
Self-Attention
}
