% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/devices.R
\name{whisper_device}
\alias{whisper_device}
\title{Device and Dtype Management}
\usage{
whisper_device()
}
\value{
torch device object
}
\description{
Utilities for managing torch devices and data types.
Get Default Device
}
