% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/devices.R
\name{whisper_dtype}
\alias{whisper_dtype}
\title{Get Default Dtype}
\usage{
whisper_dtype(device = whisper_device())
}
\arguments{
\item{device}{torch device}
}
\value{
torch dtype
}
\description{
Get Default Dtype
}
