% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/audio.R
\docType{data}
\name{WHISPER_SAMPLE_RATE}
\alias{WHISPER_SAMPLE_RATE}
\title{Audio Preprocessing for Whisper}
\format{
An object of class \code{integer} of length 1.
}
\usage{
WHISPER_SAMPLE_RATE
}
\description{
Convert audio files to mel spectrograms for Whisper input.
Whisper Audio Constants
}
\keyword{datasets}
