% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{WHISPER_SAMPLE_RATE}
\alias{WHISPER_SAMPLE_RATE}
\title{Audio Preprocessing for Whisper}
\format{
An object of class \code{integer} of length 1.
}
\usage{
WHISPER_SAMPLE_RATE
}
\description{
Convert audio files to mel spectrograms for Whisper input.
Whisper Audio Constants
}
\keyword{datasets}
